Execute bit and similar permissions problems prevent proper operation and decrease security.
Docker is not as isolated as it could otherwise be.
Different hosts will set different file bits on repository checkout.

The start.sh script, when copied on a Linux will likely not have execute bit set (which is good for security reasons).
On hosts like Windows, the execute bit might be set.

Make sure the required bits are set.
Also remove the write bit for safety reasons as well.

File permissions should probably be manually adjusted for all files COPY'd over for this reason.
This can be more complicated as some files might need to be executable and others not.

The `--from=maven` means that the permissions are being assigned from another Docker setup.
Just trust that for now.
Preliminary review suggests that there are potential permissions problems with world writable bits set on some files.
This may or may not be intentional and is therefore left alone.

Make sure `addgroup` doesn't have password generated just like with `adduser`.

The files COPY'd over are copied as the source user and the `chmod` and similar commands will probably fail.
Relocate the directory setup commands to under the root user to ensure access is allowed.
Then properly correct all permissions and switch to the non-root user.

Move all of the manipulate commands into a single RUN for a hopefully cleaner design.
Simplify the `mkdir` command by adding `-p` which results in automatically creating parent directories if they are missing.

The `/start.sh` does not need to be owned by the `$USER_NAME` for execution.
This makes the process more secure in that some process cannot alter the start script during the runtime of the container.
This prevents the user from altering the image so that docker restarts/re-runs do not initiate a modified start script.